Add files via upload
The Open Source Static Websites (OSSW) is coded by SecureSpark and with code collaboration with W3Schools! 
This Static Code Website is only going to get better. We are only offering a small portion. As that's all we can provide right now. As the project gets bigger & bigger we will continue to make the website more fancy and finish it.